Fixed the warning that appear when compiling with -Wall, which were quite annoying.
All folders, except libminisip, compile cleanly when using the most
common configuration (enable debug, enable memdebug, enable color terminal,
enable alsa). Other options may cause some un-fixed warnings.
Libminisip i did not touch as sometime all changes from the minisip
folder need to be backported ... no need to do the job twice.




git-svn-id: svn://svn.minisip.org/minisip/trunk@1511 63dae60f-c5d3-0310-92a3-fcbf00261c72

+++ libmsip/source/SipUtils.cxx
@@ -59,7 +59,7 @@ int SipUtils::findEndOfHeader(const string &buf, int &startIndex){
 				//         return
 
 	i=startIndex;
-	while (i<(int)bufSize && (buf[i]=='\r' || buf[i]=='\n' || buf[i]==' '))
+	while (i<bufSize && (buf[i]=='\r' || buf[i]=='\n' || buf[i]==' '))
 		i++;
 	startIndex = i;
